+++ core/components/minifyx/src/MinifyX.php
@@ -51,7 +51,7 @@ class MinifyX
      * The version
      * @var string $version
      */
-    public string $version = '2.0.0';
+    public string $version = '2.0.1';
 
     /**
      * The class options