+++ core/components/minifyx/docs/changelog.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2022-08-03
+
+### Fixed
+
+- Fix not set placeholders by the MinifyX snippet 
+- Fix not used minifyJs/minifyCss options by the MinifyX snippet 
+
+### Removed
+
+- Remove not used cacheFolder/cacheUrl properties from the MinifyX snippet 
+
 ## [2.0.0] - 2022-06-21
 
 ### Changed